Refactor code
Specific imports, classes, conventions
