+++ src/main/java/com/meli/socialmeli/controllers/ProductController.java
@@ -1,13 +1,23 @@
 package com.meli.socialmeli.controllers;
 
-import com.meli.socialmeli.dtos.response.PostDTO;
+import com.meli.socialmeli.dtos.request.PostDTO;
 import com.meli.socialmeli.dtos.response.PostsFromFollowsDTO;
-import com.meli.socialmeli.dtos.response.UserFollowersDTO;
 import com.meli.socialmeli.services.IProductService;
 import com.meli.socialmeli.services.impl.ProductServiceImpl;
+
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+
+
+
+
 
 @RestController
 @RequestMapping("/products")
