+++ src/main/java/com/meli/socialmeli/dtos/response/FollowersCountDTO.java
@@ -9,7 +9,7 @@ import lombok.Setter;
 @AllArgsConstructor
 @Getter
 @Setter
-public class FollowersCountDto {
+public class FollowersCountDTO {
     private int user_id;
     private String user_name;
     private int followers_count;
