+++ src/main/java/com/meli/socialmeli/dtos/response/MessageDTO.java
@@ -1,4 +1,4 @@
-package com.meli.socialmeli.dtos;
+package com.meli.socialmeli.dtos.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -9,6 +9,6 @@ import lombok.Setter;
 @AllArgsConstructor
 @Getter
 @Setter
-public class MessageDto {
+public class MessageDTO {
     private String message;
 }
