+++ src/main/java/com/meli/socialmeli/repositories/impl/UserRepositoryImpl.java
@@ -1,18 +1,20 @@
 package com.meli.socialmeli.repositories.impl;
 
-import com.meli.socialmeli.entities.User;
+
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+
+import com.meli.socialmeli.entities.User;
 import com.meli.socialmeli.exceptions.custom.DataSourceException;
 import com.meli.socialmeli.repositories.IUserRepository;
+
 import org.springframework.stereotype.Repository;
 import org.springframework.util.ResourceUtils;
 
 import java.io.File;
 import java.io.IOException;
-import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
