+++ src/main/java/com/meli/socialmeli/services/IProductService.java
@@ -1,13 +1,13 @@
 package com.meli.socialmeli.services;
 
-import com.meli.socialmeli.dtos.MessageDto;
-import com.meli.socialmeli.dtos.response.PostDTO;
+import com.meli.socialmeli.dtos.response.MessageDTO;
+import com.meli.socialmeli.dtos.request.PostDTO;
 import com.meli.socialmeli.dtos.response.PostsFromFollowsDTO;
 
 
 public interface IProductService {
 
-    MessageDto newPost(PostDTO post);
+    MessageDTO newPost(PostDTO post);
 
     PostsFromFollowsDTO getAllPostsFollowsLastTwoWeeks(Integer userId, String order);
