+++ src/main/java/com/meli/socialmeli/services/IUserService.java
@@ -1,7 +1,11 @@
 package com.meli.socialmeli.services;
 
-import com.meli.socialmeli.dtos.MessageDto;
-import com.meli.socialmeli.dtos.response.*;
+import com.meli.socialmeli.dtos.response.MessageDTO;
+import com.meli.socialmeli.dtos.response.FollowersCountDTO;
+import com.meli.socialmeli.dtos.response.UserFollowersDTO;
+import com.meli.socialmeli.dtos.response.UserResponseDTO;
+import com.meli.socialmeli.dtos.response.UserFollowedDTO;
+import com.meli.socialmeli.dtos.response.UserUnfollowDTO;
 import com.meli.socialmeli.entities.User;
 
 import java.util.List;
@@ -9,9 +13,9 @@ import java.util.List;
 public interface IUserService {
     List<User> findFollowsByIdProductService(int userId);
     UserFollowersDTO findFollowersById(int id, String order);
-    List<UserResponseDto> findAll();
-    MessageDto followSeller(int userId, int userIdToFollow);
-    FollowersCountDto getFollowersCount(int userId);
+    List<UserResponseDTO> findAll();
+    MessageDTO followSeller(int userId, int userIdToFollow);
+    FollowersCountDTO getFollowersCount(int userId);
     UserFollowedDTO findFollowedById(int userId, String order);
 
     UserUnfollowDTO unfollowUser(int userId, int userIdToUnfollow);
